requirements.txt
attrs==20.2.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
google-currency==1.0.9
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.1
itemloaders==1.0.3
jmespath==0.10.0
lxml==4.5.2
oauthlib==3.1.0
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dotenv==0.14.0
queuelib==1.5.0
requests==2.24.0
requests-oauthlib==1.3.0
Scrapy==2.3.0
service-identity==18.1.0
six==1.15.0
tweepy==3.9.0
Twisted==20.3.0
urllib3==1.25.10
w3lib==1.22.0
zope.interface==5.1.2
